fix(gui): recorder preview downgrades lazily, not on release
An instant downgrade made the combo visibly vanish for a
beat right before every normal lock-in. The preview now
freezes on the stashed burst chord for the same window the
lock-in uses — a normal full release never flickers, and
only a genuinely lingering hold settles the display to what
is actually held (scheduled settle, test-firable against
the injected clock). State machine split into
ChordRecorder+Machine.swift along the documented
shell/machine seam (file-size ceiling).
